requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
PyMuPDF==1.24.3
numpy==1.26.4
requests==2.31.0
streamlit==1.35.0
httpx==0.27.0groq
